% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountSpores.R
\docType{package}
\name{CountSpores}
\alias{CountSpores}
\title{CountSpores: Customized analysis pipeline to estimate spore growth}
\description{
For more information on the package function and an example analysis go to:
\href{https://github.com/kreutz-lab/CountSpores/}{CountSpores
}
}
