% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateOutputFolders.R
\name{createOutputFolders}
\alias{createOutputFolders}
\alias{CreateOutputFolders}
\title{Create output directory for each experiment}
\usage{
CreateOutputFolders(sheets)
}
\arguments{
\item{sheets}{name of excel sheet, corresponding to the experiment}
}
\value{
Directory for each experiment
}
\description{
Create output directory for each experiment
}
