% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickANOVAPvalues.R
\name{pickANOVAPvalues}
\alias{pickANOVAPvalues}
\alias{pickAnovaPvalues}
\title{Collects p-values form multiple ANOVA and arranges in readable format}
\usage{
pickAnovaPvalues(anov, refnames)
}
\arguments{
\item{anov}{result from ANOVA}

\item{refnames}{name of reference level used in ANOVA}
}
\value{
panova symmetric table with p-values from multiple ANOVA
}
\description{
Collects p-values form multiple ANOVA and arranges in readable format
}
