% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickTimePvalues.R
\name{pickTimePvalues}
\alias{pickTimePvalues}
\title{Collects estimated effects for multiple experiments and timepoints}
\usage{
pickTimePvalues(fits, refnames, times)
}
\arguments{
\item{fits}{results of linear mixed effects model}

\item{refnames}{name of reference level used in lme}

\item{times}{vector with timepoints of experiments}
}
\value{
estTbl seTbl pTbl estimated effects, standard errors and p-values for all timepoints in each experiment
}
\description{
Collects estimated effects for multiple experiments and timepoints
}
