% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMeanSE.R
\name{plotMeanSE}
\alias{plotMeanSE}
\title{plot mean standard error relationship for each experiment}
\usage{
plotMeanSE(file = "Data_analysis/data_proportional.xlsx")
}
\arguments{
\item{file}{path to data proportional}
}
\value{
saves png file for each experiment
}
\description{
plot mean standard error relationship for each experiment
}
