% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeCourse.R
\name{plotTimeCourse}
\alias{plotTimeCourse}
\title{plot raw data over time for each experiment}
\usage{
plotTimeCourse(file = "Data_analysis/data_proportional.xlsx")
}
\arguments{
\item{file}{path to data_proportional}
}
\value{
'sheetname'_timecourse.png in 'sheetname' folder
}
\description{
plot raw data over time for each experiment
}
