% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadData.R
\name{ReadData}
\alias{ReadData}
\title{reads in raw data and convert to useable format for analysis}
\usage{
ReadData(file, sheets)
}
\arguments{
\item{file}{path to raw data file}

\item{sheets}{character vector with sheetnames of raw data file}
}
\value{
Data_analysis  folder to store converted data files in excel format

data_proportional.xlsx columns with proportions, mean, sd, se, plotID added to raw data

data_prop_long.xlsx melted version of data_proportional
}
\description{
reads in raw data and convert to useable format for analysis
}
