% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runANOVA.R
\name{runANOVA}
\alias{runANOVA}
\title{fits lme-model to data, loops over all reference levels and runs ANOVA}
\usage{
runANOVA(file, sheets)
}
\arguments{
\item{file}{path to raw data file}

\item{sheets}{character vector with sheetnames of raw data file}
}
\value{
Data_analysis  folder to store converted data files in excel format

RandomEffectSummary.txt summarized output for effect, se and p-value for each condition and timepoint

ANOVAsummary.txt p-values for each condition, testing for effect on all timepoints
}
\description{
fits lme-model to data, loops over all reference levels and runs ANOVA
}
