% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCountSpores.R
\name{runCountSpores}
\alias{runCountSpores}
\title{Run complete CountSpores pipeline}
\usage{
runCountSpores(file)
}
\arguments{
\item{file}{path to raw data}
}
\value{
ReadData_out Output of ReadData

runAnova_out Output of runANOVA

One directory for each experiment with QC plots and test results
}
\description{
Run complete CountSpores pipeline
}
