% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMeanSE.R
\name{plotMeanSE}
\alias{plotMeanSE}
\title{plot mean standard error relationship for each experiment}
\usage{
plotMeanSE(file, sheetnames)
}
\arguments{
\item{file}{path to raw data file}

\item{sheetsnames}{character vector with sheetnames of raw data file}
}
\value{
saves png file for each experiment
}
\description{
plot mean standard error relationship for each experiment
}
