% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeCourse.R
\name{plotTimeCourse}
\alias{plotTimeCourse}
\title{plot raw data over time for each experiment}
\usage{
plotTimeCourse(file, sheetnames)
}
\arguments{
\item{file}{path to raw data file}

\item{sheetnames}{character vector with sheetnames of raw data file}
}
\value{
'sheetname'_timecourse.png in 'sheetname' folder
}
\description{
plot raw data over time for each experiment
}
