% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runANOVA.R
\name{runANOVA}
\alias{runANOVA}
\title{fits lme-model to data, loops over all possible reference levels and runs ANOVA}
\usage{
runANOVA(file = "Data_analysis/data_prop_long.xlsx")
}
\arguments{
\item{file}{path to data_prop_long}
}
\value{
Data_analysis  folder to store converted data files in excel format

results.txt Output of lme-model and ANOVA for all possible reference levels

summary_results.xlsx Excel file with complete test results and summarized views:

summary_results.xlsx sheet:LME-Model Complete output of lme-model

summary_results.xlsx sheet: ANOVA Complete output of ANOVA

summary_results.xlsx sheet: LME-Model_by_time Summary on Estimate, SE and pvalue for each timepoint

summary_results.xlsx sheet: ANOVA_summary Summarized ANOVA results based on 'Condition'

returns list of lists: aggregating for each experiment the result objects per reference level for lme-model & ANOVA
}
\description{
fits lme-model to data, loops over all possible reference levels and runs ANOVA
}
